- Made cplreader docu use the same styles like all other developer and user   docus. - cplreader docu will just be built, if CPLREADER=ON in CMake since the   plethora of commandline options to cplreader get defined in   source/cplreader/cplreader_options.cmake which in turn generates a .tex   source file which gets included into cplreader_docu.tex. - Also cplreader documentation will be built in the binary directory. - New documentation should from now on only be added to the FESpace   branch. - Xerces has been switched from 2.8.0 to 3.1.1.

+++ source/Elements/fespaceH1.cc
@@ -423,7 +423,7 @@ namespace CoupledField {
         // loop over all entities
         UInt pos = 0;
         for( UInt i = 0; i < entNumbers.GetSize(); ++i ) {
-          std::cout << "\t\t#" << std::abs(entNumbers[i]) << "\t\t";
+          std::cout << "\t\t#" << std::abs(static_cast<Double>(entNumbers[i])) << "\t\t";
 
           // leave, virtual node numbers are assigned
           for( UInt j = 0; j < offset[i]; ++j ) {
